% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/framework.R
\docType{data}
\name{framework}
\alias{framework}
\title{This is data to be included in my package}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 36 rows and 16 columns.
}
\usage{
framework
}
\description{
This is data to be included in my package
}
\references{
\url{data_blah.com}
}
\author{
My Name \email{blahblah@roxygen.org}
}
\keyword{datasets}
